[Refactor] Entity 변환 로직 타입의 한 파일로 옮김 - 사용성을 위해 해당 타입과 같은 파일 내부로 위치 수정
