+++ BoxOffice/Sources/Data/Network/DTO/DailyBoxOffice/BoxOfficeResult.swift
@@ -18,3 +18,11 @@ struct BoxOfficeResult: Decodable {
         case dailyBoxOfficeList
     }
 }
+
+// MARK: - Domain 레이어의 Entity로 변환 로직
+
+extension BoxOfficeResult {
+    func toDomain() -> [BoxOfficeEntity] {
+        return dailyBoxOfficeList.map { $0.toDomain() }
+    }
+}
